Add canned replies to browser extension (#9526)
* Add canned replies to browser extension

Move to Stack Overflow - Used when a general question is posted as an issue. Encourage the OP to post to SO instead.
Showcase submission - Used when a PR is opened to add an app to the showcase but the infoLink does not follow the guidelines.

* Update browser-extension.json

Use shorter strings to identify the canned replies.